完成角色列表，包括角色权限设置。

+++ src/views/user/role.vue
@@ -1,32 +1,16 @@
 <template>
   <div class="app-container">
     <div class="filter-container">
-      <el-input v-model="listQuery.title" :placeholder="$t('table.title')" style="width: 200px;" class="filter-item" @keyup.enter.native="handleFilter" />
-      <!-- <el-select v-model="listQuery.importance" :placeholder="$t('table.importance')" clearable style="width: 90px" class="filter-item">
-        <el-option v-for="item in importanceOptions" :key="item" :label="item" :value="item" />
-      </el-select>
-      <el-select v-model="listQuery.type" :placeholder="$t('table.type')" clearable class="filter-item" style="width: 130px">
-        <el-option v-for="item in calendarTypeOptions" :key="item.key" :label="item.display_name+'('+item.key+')'" :value="item.key" />
-      </el-select> -->
-      <!-- <el-select v-model="listQuery.sort" style="width: 140px" class="filter-item" @change="handleFilter">
-        <el-option v-for="item in sortOptions" :key="item.key" :label="item.label" :value="item.key" />
-      </el-select> -->
+      <!-- <el-input v-model="listQuery.name" placeholder="名称" style="width: 200px;" class="filter-item" @keyup.enter.native="handleFilter" />
       <el-button v-waves class="filter-item" type="primary" icon="el-icon-search" @click="handleFilter">
         {{ $t('table.search') }}
-      </el-button>
+      </el-button> -->
       <el-button class="filter-item" style="margin-left: 10px;" type="primary" icon="el-icon-edit" @click="handleCreate">
         {{ $t('table.add') }}
       </el-button>
-      <el-button v-waves :loading="downloadLoading" class="filter-item" type="primary" icon="el-icon-download" @click="handleDownload">
-        {{ $t('table.export') }}
-      </el-button>
-      <el-checkbox v-model="showReviewer" class="filter-item" style="margin-left:15px;" @change="tableKey=tableKey+1">
-        {{ $t('table.reviewer') }}
-      </el-checkbox>
     </div>
 
     <el-table
-      :key="tableKey"
       v-loading="listLoading"
       :data="list"
       border
@@ -35,20 +19,23 @@
       style="width: 100%;"
       @sort-change="sortChange"
     >
-      <el-table-column :label="$t('table.id')" prop="id" sortable="custom" align="center" width="80">
+      <el-table-column :label="$t('table.role')" align="center" width="80">
         <template slot-scope="scope">
-          <span>{{ scope.row.id }}</span>
+          <span>{{ scope.row.role }}</span>
         </template>
       </el-table-column>
-      <el-table-column :label="$t('table.title')" width="150px" align="center">
+      <el-table-column :label="$t('table.roleName')" min-width="150px" align="center">
         <template slot-scope="scope">
-          <span>{{ scope.row.role_name }}</span>
+          <span>{{ scope.row.name }}</span>
         </template>
       </el-table-column>
-      <el-table-column :label="$t('table.actions')" align="center" min-width="230" class-name="small-padding fixed-width">
+      <el-table-column :label="$t('table.actions')" align="center" width="230" class-name="small-padding fixed-width">
         <template slot-scope="{row}">
           <el-button type="primary" size="mini" @click="handleUpdate(row)">
             {{ $t('table.edit') }}
+          </el-button>
+          <el-button type="primary" size="mini" @click="handleUpdatePolicy(row)">
+            {{ $t('table.policy') }}
           </el-button>
           <el-button v-if="row.status!='deleted'" size="mini" type="danger" @click="handleModifyStatus(row,'deleted')">
             {{ $t('table.delete') }}
@@ -60,9 +47,13 @@
     <pagination v-show="total>0" :total="total" :page.sync="listQuery.page" :limit.sync="listQuery.limit" @pagination="getList" />
 
     <el-dialog :title="textMap[dialogStatus]" :visible.sync="dialogFormVisible">
-      <el-form ref="dataForm" :rules="rules" :model="temp" label-position="left" label-width="70px" style="width: 400px; margin-left:50px;">
-        <el-form-item :label="$t('table.title')" prop="title">
-          <span>{{ temp.role_name }}</span>
+      <el-form ref="dataForm" :rules="rules" :model="temp" label-position="left" label-width="90px" style="width: 400px; margin-left:50px;">
+        <el-form-item :label="$t('table.role')" prop="role">
+          <el-input v-if="dialogStatus==='create'" v-model="temp.role" />
+          <span v-else>{{ temp.role }}</span>
+        </el-form-item>
+        <el-form-item :label="$t('table.roleName')" prop="name">
+          <el-input v-model="temp.name" />
         </el-form-item>
       </el-form>
       <div slot="footer" class="dialog-footer">
@@ -75,13 +66,19 @@
       </div>
     </el-dialog>
 
-    <el-dialog :visible.sync="dialogPvVisible" title="Reading statistics">
-      <el-table :data="pvData" border fit highlight-current-row style="width: 100%">
-        <el-table-column prop="key" label="Channel" />
-        <el-table-column prop="pv" label="Pv" />
-      </el-table>
+    <el-dialog :visible.sync="dialogPolicyVisible" title="权限修改">
+      <el-checkbox v-model="checkAll" :indeterminate="isIndeterminate" @change="handleCheckAllChange">全选</el-checkbox>
+      <div style="margin: 15px 0;" />
+      <el-checkbox-group v-model="checkedPolicys" @change="handleCheckedPolicysChange">
+        <el-checkbox v-for="policy in allPolicys" :key="policy.policy" :label="policy.policy">{{ policy.name }}</el-checkbox>
+      </el-checkbox-group>
       <span slot="footer" class="dialog-footer">
-        <el-button type="primary" @click="dialogPvVisible = false">{{ $t('table.confirm') }}</el-button>
+        <el-button @click="dialogPolicyVisible = false">
+          {{ $t('table.cancel') }}
+        </el-button>
+        <el-button type="primary" @click="handleUpdateRolePolicy()">
+          {{ $t('table.confirm') }}
+        </el-button>
       </span>
     </el-dialog>
   </div>
@@ -89,8 +86,8 @@
 
 <script>
 import { getRoles, addRole, updateRole, deleteRole } from '@/api/role'
+import { getPolicyByRole, setPolicyByRole } from '@/api/policy'
 import waves from '@/directive/waves' // waves directive
-import { parseTime } from '@/utils'
 import Pagination from '@/components/Pagination' // secondary package based on el-pagination
 
 export default {
@@ -109,7 +106,6 @@ export default {
   },
   data() {
     return {
-      tableKey: 0,
       list: null,
       total: 0,
       listLoading: true,
@@ -117,7 +113,7 @@ export default {
         page: 1,
         limit: 20,
         importance: undefined,
-        title: undefined,
+        name: undefined,
         type: undefined,
         sort: '+id'
       },
@@ -135,17 +131,22 @@ export default {
         status: 'published'
       },
       dialogFormVisible: false,
+      dialogPolicyVisible: false,
       dialogStatus: '',
       textMap: {
         update: 'Edit',
         create: 'Create'
       },
-      dialogPvVisible: false,
-      pvData: [],
       rules: {
-        // title: [{ required: true, message: 'title is required', trigger: 'blur' }]
+        role: [{ required: true, message: 'role is required', trigger: 'blur' }],
+        name: [{ required: true, message: 'name is required', trigger: 'blur' }]
       },
-      downloadLoading: false
+      checkAll: false,
+      checkedPolicys: [],
+      checkedAllPolicys: [],
+      allPolicys: [],
+      isIndeterminate: false,
+      currentRole: ''
     }
   },
   created() {
@@ -161,7 +162,7 @@ export default {
         // Just to simulate the time of the request
         setTimeout(() => {
           this.listLoading = false
-        }, 1.5 * 1000)
+        }, 1 * 500)
       })
     },
     handleFilter() {
@@ -169,35 +170,19 @@ export default {
       this.getList()
     },
     handleModifyStatus(row, status) {
-      this.$message({
-        message: '操作成功',
-        type: 'success'
-      })
-      row.status = status
+      if (status === 'deleted') {
+        this.handleDelete(row)
+      }
     },
     sortChange(data) {
-      const { prop, order } = data
-      if (prop === 'id') {
-        this.sortByID(order)
-      }
     },
     sortByID(order) {
-      if (order === 'ascending') {
-        this.listQuery.sort = '+id'
-      } else {
-        this.listQuery.sort = '-id'
-      }
       this.handleFilter()
     },
     resetTemp() {
       this.temp = {
-        id: undefined,
-        importance: 1,
-        remark: '',
-        timestamp: new Date(),
-        title: '',
-        status: 'published',
-        type: ''
+        role: '',
+        name: ''
       }
     },
     handleCreate() {
@@ -211,8 +196,7 @@ export default {
     createData() {
       this.$refs['dataForm'].validate((valid) => {
         if (valid) {
-          this.temp.id = parseInt(Math.random() * 100) + 1024 // mock a id
-          this.temp.author = 'vue-element-admin'
+          this.temp.author = 'vue-element-admin'// TODO
           addRole(this.temp).then(() => {
             this.list.unshift(this.temp)
             this.dialogFormVisible = false
@@ -228,11 +212,37 @@ export default {
     },
     handleUpdate(row) {
       this.temp = Object.assign({}, row) // copy obj
-      this.temp.timestamp = new Date(this.temp.timestamp)
       this.dialogStatus = 'update'
       this.dialogFormVisible = true
       this.$nextTick(() => {
         this.$refs['dataForm'].clearValidate()
+      })
+    },
+    handleUpdatePolicy(row) {
+      // this.temp = Object.assign({}, row) // copy obj
+      this.currentRole = row.role
+      getPolicyByRole({ role: row.role }).then(response => {
+        this.allPolicys = []
+        this.checkedPolicys = []
+        this.checkedAllPolicys = []
+        response.data.all_policy_items.forEach(item => {
+          this.checkedAllPolicys.push(item.policy)
+          if (item.name !== '') {
+            this.allPolicys.push(item)
+          } else {
+            item.name = item.policy
+            this.allPolicys.push(item)
+          }
+        })
+        response.data.role_policy_items.forEach(item => {
+          this.checkedPolicys.push(item.policy)
+        })
+        if (this.checkedPolicys.length > 0) {
+          this.isIndeterminate = true
+        } else {
+          this.isIndeterminate = false
+        }
+        this.dialogPolicyVisible = true
       })
     },
     updateData() {
@@ -259,39 +269,38 @@ export default {
       })
     },
     handleDelete(row) {
-      deleteRole(this.temp).then(() => {
-        this.$notify({
-          title: '成功',
-          message: '删除成功',
-          type: 'success',
-          duration: 2000
+      this.$confirm('是否删除角色？', '确认信息', {
+        distinguishCancelAndClose: true,
+        confirmButtonText: '删除',
+        cancelButtonText: '取消'
+      }).then(() => {
+        this.temp = Object.assign({}, row) // copy obj
+        deleteRole(this.temp).then(() => {
+          const index = this.list.indexOf(row)
+          this.list.splice(index, 1)
+          this.$notify({
+            title: '成功',
+            message: '删除成功',
+            type: 'success',
+            duration: 2000
+          })
         })
       })
-      const index = this.list.indexOf(row)
-      this.list.splice(index, 1)
     },
-    handleDownload() {
-      this.downloadLoading = true
-      import('@/vendor/Export2Excel').then(excel => {
-        const tHeader = ['timestamp', 'title', 'type', 'importance', 'status']
-        const filterVal = ['timestamp', 'title', 'type', 'importance', 'status']
-        const data = this.formatJson(filterVal, this.list)
-        excel.export_json_to_excel({
-          header: tHeader,
-          data,
-          filename: 'table-list'
-        })
-        this.downloadLoading = false
+    handleCheckAllChange(val) {
+      this.checkedPolicys = val ? this.checkedAllPolicys : []
+      this.isIndeterminate = false
+    },
+    handleCheckedPolicysChange(value) {
+      const checkedCount = value.length
+      this.checkAll = checkedCount === this.allPolicys.length
+      this.isIndeterminate = checkedCount > 0 && checkedCount < this.allPolicys.length
+    },
+    handleUpdateRolePolicy() {
+      setPolicyByRole({ role: this.currentRole, policys: this.checkedPolicys }).then(() => {
+        this.currentRole = ''
+        this.dialogPolicyVisible = false
       })
-    },
-    formatJson(filterVal, jsonData) {
-      return jsonData.map(v => filterVal.map(j => {
-        if (j === 'timestamp') {
-          return parseTime(v[j])
-        } else {
-          return v[j]
-        }
-      }))
     }
   }
 }
